Leverage: slippage calculation improvements (#644)
- Use a single useQuery() for the two requests
- useCheckLeverageSlippage() now takes collIndex a parameter rather than a loan object
- Move useCheckLeverageSlippage() to liquity-leverage.ts
